@@ -1,10 +1,15 @@
 package searchengine.controllers;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import searchengine.dto.statistics.ApiResponse;
 import searchengine.dto.statistics.StatisticsResponse;
+import searchengine.exceptions.IndexingStartException;
+import searchengine.services.CrawlerService;
+import searchengine.services.IndexingService;
 import searchengine.services.StatisticsService;
 
 @RestController
@@ -12,13 +17,43 @@ import searchengine.services.StatisticsService;
 public class ApiController {
 
     private final StatisticsService statisticsService;
+    private final IndexingService indexingService;
 
-    public ApiController(StatisticsService statisticsService) {
+    public ApiController(StatisticsService statisticsService, IndexingService indexingService) {
         this.statisticsService = statisticsService;
+        this.indexingService = indexingService;
     }
 
     @GetMapping("/statistics")
     public ResponseEntity<StatisticsResponse> statistics() {
         return ResponseEntity.ok(statisticsService.getStatistics());
+    }
+
+    @GetMapping("/startIndexing")
+    public ResponseEntity<ApiResponse> startIndexing() {
+        try {
+            indexingService.startIndexing();
+        } catch (IndexingStartException e) {
+            return ResponseEntity.status(HttpStatus.CONFLICT)
+                    .body(
+                            ApiResponse.error(e.getMessage())
+                    );
+        }
+
+        return ResponseEntity.ok(ApiResponse.ok());
+    }
+
+    @GetMapping("/stopIndexing")
+    public ResponseEntity<ApiResponse> stopIndexing() {
+        try {
+            indexingService.stopIndexing();
+        } catch (IndexingStartException e) {
+            return ResponseEntity.status(HttpStatus.CONFLICT)
+                    .body(
+                            ApiResponse.error(e.getMessage())
+                    );
+        }
+
+        return ResponseEntity.ok(ApiResponse.ok());
     }
 }
